refactor(composio): split ops.rs (1746 lines) into ops/ directory
10 focused sub-modules: error_utils, toolkits, connections,
memory_cleanup, tools_ops, execute, triggers, providers_ops,
direct_mode. All pub(crate) visibility preserved.
